chore(react): move Social to sds and re-section Timeline
Aligns folder location with Storybook section for the two remaining
sds/kits mismatches:

- Social (Reactions) moves from src/kits/ to src/sds/social and now
  shows under "Domain specific/social" instead of "Components".
- Timeline (already physically in src/sds/) moves in Storybook from
  "Patterns/Timeline" to "Domain specific/timeline".

Per review feedback, the sds group folders are lowercased (social,
timeline) to match the existing ai/surveys/chat/inbox convention;
capitalized names stay reserved for components (Reactions, TimelineRow).

Reactions keeps flowing through the experimental entry via the existing
@deprecated re-export (repointed), so consumers need no migration.
