+++ packages/react/src/components/exports.ts
@@ -66,7 +66,7 @@ export * from "./F0Select"
 export * from "./F0Slider"
 export * from "./F0Text"
 export * from "./F0VideoPlayer"
-export * from "../sds/TimeLine"
+export * from "../sds/timeline"
 /**
  * @deprecated OneFilterPicker has moved to @/patterns/OneFilterPicker. Import from there instead.
  */
