+++ packages/react/src/kits/surveys/SurveyFormBuilder/QuestionTypes/BaseScoreQuestion/ScoreEditOption.tsx
@@ -2,7 +2,7 @@ import data from "@emoji-mart/data/sets/15/twitter.json"
 import { useEffect, useState } from "react"
 
 import { F0Button } from "@/components/F0Button"
-import "@/kits/Social/Reactions/Picker/index.css"
+import "@/sds/social/Reactions/Picker/index.css"
 import { EmojiPicker } from "@/lib/EmojiPicker"
 import { cn } from "@/lib/utils"
 import { Popover, PopoverContent, PopoverTrigger } from "@/ui/popover"
