+++ packages/react/src/sds/Home/Communities/Post/CommunityPost/index.tsx
@@ -3,7 +3,7 @@ import { F0AvatarPerson } from "@/components/avatars/F0AvatarPerson"
 import { F0Button } from "@/components/F0Button"
 import { IconType } from "@/components/F0Icon"
 import { F0Link } from "@/components/F0Link"
-import { Reactions, ReactionsProps } from "@/kits/Social/Reactions"
+import { Reactions, ReactionsProps } from "@/sds/social/Reactions"
 import { Dropdown, DropdownItem } from "@/experimental/Navigation/Dropdown"
 import {
   Comment as CommentIcon,
