+++ packages/react/src/sds/social/Reactions/reaction.tsx
