+++ packages/react/src/sds/timeline/types.ts
